requirements.txt
jinja2
pymongo
python-dateutil
simplejson
cherrypy
python-memcached